requirements.txt
google-genai
openai
requests
redis
wheel==0.45.1
replicate
Pillow